+++ controllers/thoughtController.js
@@ -92,12 +92,12 @@ module.exports = {
       res.status(500).json(err);
     }
   },
-  // Add a thought response
-  async addThoughtResponse(req, res) {
+  // Add a thought reaction
+  async addReaction(req, res) {
     try {
       const thought = await Thought.findOneAndUpdate(
         { _id: req.params.thoughtId },
-        { $addToSet: { responses: req.body } },
+        { $addToSet: { reactions: req.body } },
         { runValidators: true, new: true }
       );
 
@@ -110,12 +110,12 @@ module.exports = {
       res.status(500).json(err);
     }
   },
-  // Remove thought response
-  async removeThoughtResponse(req, res) {
+  // Remove thought reaction
+  async removeThoughtReaction(req, res) {
     try {
       const thought = await Thought.findOneAndUpdate(
         { _id: req.params.thoughtId },
-        { $pull: { reactions: { responseId: req.params.responseId } } },
+        { $pull: { reactions: { reactiond: req.params.reactionId } } },
         { runValidators: true, new: true }
       );