+++ routes/api/thoughtRoutes.js
@@ -18,7 +18,7 @@ router.route('/').get(getThoughts).post(createThought);
 router.route('/:thoughtId').get(getSingleThought).put(updateThought).delete(deleteThought);
 
 // /api/thoughts/:thoughtId/reactions
-// router.route('/:thoughtId/reactions').post(addReaction);
+router.route('/:thoughtId/reactions').post(addReaction);
 // router.route('/:thoughtId/reactions/:reactionId').delete(removeReaction)
 
 module.exports = router;