+++ CHANGELOG.md
@@ -10,3 +10,4 @@
 8. Updated User Route & userController to add Thought and update user
 9. Updated User Route & userController to add a Friend
 10. Updated User Route & userController to remove a Friend
+11. Updated Thougght Route and thoughtController to add a Reaction